Promote Slider and RangeSlider to core
