+++ .storybook/decorators/withResponsiveWrapper.tsx
@@ -1,5 +1,10 @@
-import { ToggleButton, ToggleButtonGroup, Tooltip } from "@salt-ds/core";
-import { NumberInput, Slider } from "@salt-ds/lab";
+import {
+  Slider,
+  ToggleButton,
+  ToggleButtonGroup,
+  Tooltip,
+} from "@salt-ds/core";
+import { NumberInput } from "@salt-ds/lab";
 import type { Decorator } from "@storybook/react";
 import { type ReactNode, type SyntheticEvent, useState } from "react";
 import "./ResponsiveContainer.css";
