+++ packages/core/src/__tests__/__e2e__/slider/RangeSlider.cy.tsx
