+++ packages/core/src/slider/index.ts
