+++ packages/core/src/slider/internal/SliderThumb.css
