+++ packages/core/src/slider/internal/SliderTooltip.tsx
@@ -1,7 +1,8 @@
-import { Text, makePrefixer } from "@salt-ds/core";
 import { useComponentCssInjection } from "@salt-ds/styles";
 import { useWindow } from "@salt-ds/window";
 import clsx from "clsx";
+import { Text } from "../../text";
+import { makePrefixer } from "../../utils";
 import sliderTooltipCss from "./SliderTooltip.css";
 
 const withBaseName = makePrefixer("saltSliderTooltip");
