+++ packages/core/src/slider/internal/SliderTrack.css
