+++ packages/core/src/slider/internal/useRangeSliderThumb.ts
