+++ packages/core/src/slider/internal/useSliderThumb.ts
