+++ packages/core/src/slider/internal/utils.ts
