+++ site/docs/components/slider/index.mdx
@@ -2,9 +2,9 @@
 title: Slider
 data:
   description: "`Slider` enables the selection of a single value within a range of values. It offers inline label positioning and marks displayed under the track."
-  sourceCodeUrl: "https://github.com/jpmorganchase/salt-ds/tree/main/packages/lab/src/slider"
+  sourceCodeUrl: "https://github.com/jpmorganchase/salt-ds/tree/main/packages/core/src/slider"
   package:
-    name: "@salt-ds/lab"
+    name: "@salt-ds/core"
   alsoKnownAs: []
   relatedComponents: []
 layout: DetailComponent
