+++ site/docs/components/slider/range-slider/index.mdx
@@ -2,7 +2,7 @@
 title: Range slider
 data:
   description: "`RangeSlider` enables the selection of an interval within a range of values. It offers inline label positioning and marks displayed under the track."
-  sourceCodeUrl: "https://github.com/jpmorganchase/salt-ds/tree/main/packages/lab/src/range-slider"
+  sourceCodeUrl: "https://github.com/jpmorganchase/salt-ds/tree/main/packages/core/src/slider"
   alsoKnownAs: []
   relatedComponents: []
 layout: DetailComponent
