+++ site/docs/components/slider/range-slider/usage.mdx
@@ -23,12 +23,12 @@ data:
 
 ## Import
 
-To import `RangeSlider` from the lab Salt package, use:
+To import `RangeSlider` from the core Salt package, use:
 
 ```js
-import { RangeSlider } from "@salt-ds/lab";
+import { RangeSlider } from "@salt-ds/core";
 ```
 
 ## Props
 
-<PropsTable packageName="lab" componentName="RangeSlider" />
+<PropsTable packageName="core" componentName="RangeSlider" />
