+++ site/docs/components/slider/usage.mdx
@@ -24,12 +24,12 @@ data:
 
 ## Import
 
-To import `Slider` from the lab Salt package, use:
+To import `Slider` from the core Salt package, use:
 
 ```js
-import { Slider } from "@salt-ds/lab";
+import { Slider } from "@salt-ds/core";
 ```
 
 ## Props
 
-<PropsTable packageName="lab" componentName="Slider" />
+<PropsTable packageName="core" componentName="Slider" />
