+++ site/src/examples/range-slider/WithLabels.tsx
@@ -1,5 +1,9 @@
-import { FormField, FormFieldLabel, StackLayout } from "@salt-ds/core";
-import { RangeSlider } from "@salt-ds/lab";
+import {
+  FormField,
+  FormFieldLabel,
+  RangeSlider,
+  StackLayout,
+} from "@salt-ds/core";
 import type { ReactElement } from "react";
 
 export const WithLabels = (): ReactElement => (
