+++ site/src/examples/range-slider/WithNonNumericValues.tsx
@@ -1,5 +1,4 @@
-import { useResponsiveProp } from "@salt-ds/core";
-import { RangeSlider } from "@salt-ds/lab";
+import { RangeSlider, useResponsiveProp } from "@salt-ds/core";
 import { type ReactElement, useState } from "react";
 
 export const WithNonNumericValues = (): ReactElement => {
