+++ site/src/examples/slider/WithCustomStep.tsx
@@ -1,5 +1,4 @@
-import { FormField, FormFieldLabel, StackLayout } from "@salt-ds/core";
-import { Slider } from "@salt-ds/lab";
+import { FormField, FormFieldLabel, Slider, StackLayout } from "@salt-ds/core";
 import type { ReactElement } from "react";
 
 export const WithCustomStep = (): ReactElement => (
