+++ site/src/examples/slider/WithMarks.tsx
@@ -1,5 +1,4 @@
-import { FormField, FormFieldLabel, StackLayout } from "@salt-ds/core";
-import { Slider } from "@salt-ds/lab";
+import { FormField, FormFieldLabel, Slider, StackLayout } from "@salt-ds/core";
 import type { ReactElement } from "react";
 
 export const WithMarks = (): ReactElement => (
@@ -83,6 +82,7 @@ export const WithMarks = (): ReactElement => (
         maxLabel="High"
         accessibleMinText="Low"
         accessibleMaxText="High"
+        showTicks
         marks={[
           {
             value: 0,
