+++ site/src/examples/slider/WithNonNumericValues.tsx
@@ -1,5 +1,4 @@
-import { useResponsiveProp } from "@salt-ds/core";
-import { Slider } from "@salt-ds/lab";
+import { Slider, useResponsiveProp } from "@salt-ds/core";
 import { type ReactElement, useState } from "react";
 
 export const WithNonNumericValues = (): ReactElement => {
